src/ML/requirements.txt
gunicorn
flask
pandas
numpy
tensorflow
scikit-learn
joblib
google-cloud-bigquery
db-dtypes
pandas_gbq